🚚 Move: 파일 세분화 작업
