+++ PADO/PADO/iOS/Setting/View/Components/SettingDevider.swift
@@ -1,0 +1,25 @@
+//
+//  SettingProfileDevider.swift
+//  PADO
+//
+//  Created by 황민채 on 1/16/24.
+//
+import SwiftUI
+
+struct SettingProfileDevider: View {
+    var body: some View {
+        Rectangle()
+            .frame(width: UIScreen.main.bounds.width * 0.9, height: 0.7)
+            .foregroundStyle(.gray)
+            .opacity(0.3)
+    }
+}
+
+struct SettingDevider: View {
+    var body: some View {
+        Rectangle()
+            .frame(width: UIScreen.main.bounds.width * 0.9, height: 0.3)
+            .opacity(0.4)
+            .foregroundStyle(.gray)
+    }
+}
